requirements.txt
streamlit
python-docx
PyMuPDF
pandas

